refactor: Extract converters and utilities for Google GenAI client integration
- Introduced `GoogleGenaiConversionUtils` for JSON/map conversion, signature encoding, and SDK schema handling.
- Added `GoogleGenaiToolConverter` for mapping `ToolDescriptor` to Google GenAI SDK types.
- Implemented `GoogleGenaiRequestConverter` and `GoogleGenaiResponseConverter` to simplify request and response handling.
- Refactored `GoogleGenaiLLMClient` to use modular converters for improved maintainability and readability.
